fix the conflict between autogptq and sglang
